requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pillow==10.2.0
torch==2.2.2+cpu
torchvision==0.17.2+cpu
facenet-pytorch==2.6.0
scikit-learn==1.6.1
numpy==1.26.4
python-dotenv==1.0.0